requirements.txt
openai>=1.30.0
google-genai>=1.0.0
requests>=2.31.0
Pillow>=10.0.0
customtkinter>=5.2.0
